[IMP] web: convert error_service tests to hoot
task-id: 3705027
